テストケースを追加．

+++ Graphics/Polygon.cpp
@@ -7,7 +7,7 @@
 using namespace Crystal::Math;
 using namespace Crystal::Graphics;
 
-void Polygon::rotateX(const float angle)
+void Polygon::rotateX(const double angle)
 {
 	move(-1.0 *center);
 	const Matrix3d<double>& m = Matrix3d<double>::RotateX(angle * Tolerances::getPrecisePI() / 180.0);
@@ -17,7 +17,7 @@ void Polygon::rotateX(const float angle)
 	move(1.0 * center);
 }
 
-void Polygon::rotateY(const float angle)
+void Polygon::rotateY(const double angle)
 {
 	move(-1.0 *center);
 	const Matrix3d<double>& m = Matrix3d<double>::RotateY(angle * Tolerances::getPrecisePI() / 180.0);
@@ -27,7 +27,7 @@ void Polygon::rotateY(const float angle)
 	move(1.0 * center);
 }
 
-void Polygon::rotateZ(const float angle)
+void Polygon::rotateZ(const double angle)
 {
 	move(-1.0 *center);
 	const Matrix3d<double>& m = Matrix3d<double>::RotateZ(angle * Tolerances::getPrecisePI()/ 180.0);
